Clean up test infrastructure: remove dead fixtures, deduplicate DataFrames, fix imports
- Fix unused Mock imports in test_chunked_enricher.py and
  test_chunked_enrichment.py (Copilot review feedback)
- Consolidate duplicate sample_sbir_df and sample_recipient_df fixtures
  from two test files into enrichers/conftest.py (were identical copies)
- Remove unused cached_test_companies and cached_test_awards session
  fixtures from root conftest (never referenced by any test)
- Remove orphaned tests/unit/fiscal/conftest.py (directory has no tests)
- Remove 4 unused re-exports from transition/conftest.py
  (default_transition_config, sample_contracts_df, recent_award, old_award)
- Remove 3 empty/pass-only test classes from test_exception_handling.py
  (TestAPIErrorHandling, TestValidationErrorHandling, TestDependencyErrorHandling)
- Drop unused mock_enrichment_config re-export from enrichers/conftest.py
