+++ tests/unit/enrichers/conftest.py
@@ -1,11 +1,27 @@
-"""Shared fixtures for enricher tests.
+"""Shared fixtures for enricher tests."""
 
-Import fixtures from the shared conftest module.
-"""
-
-# Import shared fixtures
-from tests.conftest_shared import mock_enrichment_config, sample_recipient_df, sample_sbir_df
+import pandas as pd
+import pytest
 
 
-# Re-export for pytest discovery
-__all__ = ["mock_enrichment_config", "sample_sbir_df", "sample_recipient_df"]
+@pytest.fixture
+def sample_sbir_df():
+    """Sample SBIR DataFrame for enricher tests."""
+    return pd.DataFrame(
+        {
+            "award_id": [f"AWD-{i}" for i in range(5)],
+            "company_name": [f"Company {i}" for i in range(5)],
+            "company_uei": [f"UEI{i:09d}" for i in range(5)],
+        }
+    )
+
+
+@pytest.fixture
+def sample_recipient_df():
+    """Sample recipient DataFrame for enricher tests."""
+    return pd.DataFrame(
+        {
+            "recipient_name": ["Acme Corp", "TechStart Inc"],
+            "recipient_uei": ["UEI000000001", "UEI000000002"],
+        }
+    )
