+++ tests/unit/enrichers/test_chunked_enricher.py
@@ -1,8 +1,7 @@
 """Tests for ChunkedEnricher initialization and core functionality."""
 
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
-import pandas as pd
 import pytest
 
 from sbir_etl.enrichers.chunked_enrichment import ChunkedEnricher
@@ -16,29 +15,6 @@ pytestmark = pytest.mark.fast
 def mock_config():
     """Mock configuration for testing."""
     return create_mock_enrichment_performance_config()
-
-
-@pytest.fixture
-def sample_sbir_df():
-    """Sample SBIR DataFrame."""
-    return pd.DataFrame(
-        {
-            "award_id": [f"AWD-{i}" for i in range(5)],
-            "company_name": [f"Company {i}" for i in range(5)],
-            "company_uei": [f"UEI{i:09d}" for i in range(5)],
-        }
-    )
-
-
-@pytest.fixture
-def sample_recipient_df():
-    """Sample recipient DataFrame."""
-    return pd.DataFrame(
-        {
-            "recipient_name": ["Acme Corp", "TechStart Inc"],
-            "recipient_uei": ["UEI000000001", "UEI000000002"],
-        }
-    )
 
 
 @pytest.fixture
